more rearranging, fixes
