+++ package.json
@@ -34,7 +34,6 @@
       "express": "^4.18.2",
       "express-fileupload": "^1.2.0",
       "isomorphic-git": "^1.25.6",
-      "jsonschema": "^1.2.6",
       "morgan": "^1.10.0",
       "mysql": "^2.18.1",
       "node-ssh": "^10.0.2",
@@ -43,7 +42,8 @@
       "rimraf": "^5.0.5",
       "typeorm": "0.3.20",
       "util": "^0.12.3",
-      "validator": "^13.7.0"
+      "validator": "^13.7.0",
+      "zod": "^3.24.1"
    },
    "devDependencies": {
       "@inscreen/typeorm-uml": "^1.6.5-is1",
