+++ src/definitions/Errors.ts
