+++ src/definitions/UtilTypes.ts
@@ -1,0 +1,3 @@
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type callableFunction = (..._args: any[]) => unknown;
