+++ src/helpers/Helper.ts
@@ -1,6 +1,6 @@
 import { config, hpcConfigMap, jupyterGlobusMap } from "../../configs/config";
+import { callableFunction } from "../definitions";
 import { Job } from "../models/Job";
-import { callableFunction } from "../utils/types";
 // import * as fs from "fs";
