+++ src/maintainers/CommunityContributionMaintainer.ts
@@ -1,4 +1,5 @@
 import SingularityConnector from "../connectors/SingularityConnector";
+import { executableManifest, GitFolder } from "../definitions";
 import GitUtil from "../helpers/GitUtil";
 import * as Helper from "../helpers/Helper";
 import XSEDEUtil from "../helpers/XSEDEUtil";
@@ -7,7 +8,6 @@ import { Git } from "../models/Git";
 import dataSource from "../utils/DB";
 import { BaseFolderUploader, FolderUploaderHelper } from "../utils/FolderUploader";
 import { ResultFolderContentManager } from "../utils/Redis";
-import { executableManifest, GitFolder } from "../utils/types";
 
 import BaseMaintainer from "./BaseMaintainer";
