+++ src/utils/FolderUploader.ts
@@ -7,6 +7,14 @@ import { hpcConfigMap } from "../../configs/config";
 import BaseConnector from "../connectors/BaseConnector";
 import SingularityConnector from "../connectors/SingularityConnector";
 import SlurmConnector from "../connectors/SlurmConnector";
+import { NotImplementedError } from "../defines/Errors";
+import {
+  BaseFolder,
+  GitFolder,
+  GlobusFolder,
+  hpcConfig,
+  LocalFolder,
+} from "../definitions";
 import FolderUtil from "../helpers/FolderUtil";
 import GitUtil from "../helpers/GitUtil";
 import { GlobusClient } from "../helpers/GlobusTransferUtil";
@@ -15,14 +23,6 @@ import { Cache } from "../models/Cache";
 import { Folder } from "../models/Folder";
 
 import dataSource from "./DB";
-import { NotImplementedError } from "./errors";
-import {
-  BaseFolder,
-  GitFolder,
-  GlobusFolder,
-  hpcConfig,
-  LocalFolder,
-} from "./types";
 
 type Connector =
   | BaseConnector
