+++ src/utils/Supervisor.ts
@@ -4,6 +4,7 @@ import * as events from "events";
 
 import { config, maintainerConfigMap, hpcConfigMap } from "../../configs/config";
 import connectionPool from "../connectors/ConnectionPool";
+import { SSH, callableFunction } from "../definitions";
 import * as Helper from "../helpers/Helper";
 import BaseMaintainer from "../maintainers/BaseMaintainer";
 import { Job } from "../models/Job";
@@ -11,7 +12,6 @@ import { Job } from "../models/Job";
 import dataSource from "./DB";
 import Emitter from "./Emitter";
 import { JobQueue } from "./Redis";
-import { SSH, callableFunction } from "./types";
 
 /**
  * Manages
